jpaの練習
