+++ spring-jpa2/src/main/resources/application.properties
@@ -1,3 +1,1 @@
-
-spring.jpa.show-sql=true
-spring.datasource.url=jdbc:h2:mem:testdb
+spring.autoconfigure.exclude=org.springframework.boot.devtools.autoconfigure.DevToolsDataSourceAutoConfiguration
