+++ spring-jpa3/src/main/java/com/example/demo/model/Employee.java
@@ -3,11 +3,10 @@ package com.example.demo.model;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import javax.persistence.ManyToOne;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Size;
-
-import org.springframework.data.annotation.Id;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -16,16 +15,15 @@ import lombok.Setter;
 @Setter
 @Entity
 public class Employee {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)//主キーを生成
-	private Long id;
+    @NotBlank
+    @Size(max = 40)
+    private String name;
 
-	@NotBlank
-	@Size(max = 40)
-	private String name;
-
-	@ManyToOne//多対一のテーブルを指定 部署に対して 1:多
-	private Department department;
-
+    @ManyToOne
+    private Department department;
 }
+
