+++ spring-jpa3/src/main/java/com/example/demo/repository/DepartmentRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Department;
 
-public interface DepartmentRepository extends JpaRepository<Department,Long>{
-
+public interface DepartmentRepository extends JpaRepository<Department, Long> {
 }
