+++ spring-jpa3/src/main/java/com/example/demo/repository/EmployeeRepository.java
@@ -5,5 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.model.Employee;
 
 public interface EmployeeRepository extends JpaRepository<Employee, Long> {
-
 }
