+++ spring-jpa3/src/main/resources/data.sql
@@ -1,0 +1,2 @@
+insert into department(name)
+values ('営業'),('経理'),('開発'),('総務')
