+++ spring-jpa3/src/main/resources/templates/form.html
@@ -1,0 +1,70 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+  <meta charset="UTF-8">
+  <th:block th:insert="fragments/base :: header"></th:block>
+</head>
+<body id="page-top">
+  <!-- Page Wrapper -->
+  <div id="wrapper">
+    <th:block th:insert="fragments/sidebar :: sidebar"></th:block>
+
+    <!-- Main -->
+    <div id="content-wrapper" class="d-flex flex-column">
+      <div id="content">
+        <div class="container-fluid">
+
+          <!-- DataTales -->
+          <div class="card shadow my-4">
+            <div class="card-header py-3">
+              <h6 class="m-0 font-weight-bold text-primary">
+                社員登録
+              </h6>
+            </div>
+            <div class="card-body">
+              <div class="table-responsive">
+                <form th:action="@{/process}" th:object="${employee}"
+                  method="post">
+                  <input type="hidden" th:field="*{id}">
+
+                  <div class="form-group">
+                    <label for="name">名前</label>
+                    <input type="text" class="form-control"
+                      th:errorclass="is-invalid" th:field="*{name}">
+                    <div class="invalid-feedback" th:errors="*{name}">
+                    </div>
+                  </div>
+
+                  <div class="form-group">
+                    <label for="department">部署名</label>
+                    <select class="form-control" th:field="*{department}">
+                      <th:block
+                        th:each="department : ${@departmentRepository.findAll()}">
+                        <option th:value="${department.id}"
+                          th:text="${department.name}">
+                        </option>
+                      </th:block>
+                    </select>
+                  </div>
+
+                  <!-- Divider -->
+                  <hr class="sidebar-divider my-4">
+
+                  <button class="btn btn-primary btn-block col-md-4">
+                    <i class="fas fa-edit fa-fw"></i>
+                    <span class="text">保存</span>
+                  </button>
+                </form>
+              </div>
+            </div>
+          </div>
+
+        </div>
+      </div>
+    </div>
+    <!-- End of Main -->
+  </div>
+
+  <th:block th:insert="fragments/base :: scripts"></th:block>
+</body>
+</html>
