+++ spring-jpa3/src/main/resources/templates/fragments/base.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+  <meta charset="UTF-8">
+
+  <th:block th:fragment="header">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>社員情報アプリ</title>
+
+    <!-- These are required CSS -->
+    <link rel="stylesheet" th:href="@{/webjars/font-awesome/css/all.min.css}">
+    <link rel="stylesheet" th:href="@{/webjars/startbootstrap-sb-admin-2/css/sb-admin-2.min.css}">
+  </th:block>
+</head>
+<body>
+
+  <!-- These are required JS -->
+  <th:block th:fragment="scripts">
+    <script th:src="@{/webjars/jquery/jquery.min.js}"></script>
+    <script th:src="@{/webjars/popper.js/umd/popper.min.js}"></script>
+    <script th:src="@{/webjars/bootstrap/js/bootstrap.min.js}"></script>
+    <script th:src="@{/webjars/startbootstrap-sb-admin-2/js/sb-admin-2.min.js}"></script>
+    <script th:src="@{/webjars/chartjs/Chart.min.js}"></script>
+  </th:block>
+</body>
+</html>
