+++ spring-jpa3/src/main/resources/templates/fragments/sidebar.html
@@ -1,0 +1,54 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+  <meta charset="UTF-8">
+</head>
+<body>
+  <th:block th:fragment="sidebar">
+    <!-- Sidebar -->
+    <ul class="navbar-nav bg-gradient-primary sidebar sidebar-dark accordion"
+      id="accordionSidebar">
+
+      <!-- Sidebar - Brand -->
+      <a class="sidebar-brand d-flex align-items-center justify-content-center"
+        th:href="@{/}">
+        <div class="sidebar-brand-icon rotate-n-15">
+          <i class="fas fa-id-card"></i>
+        </div>
+        <div class="sidebar-brand-text mx-3">
+          社員情報
+        </div>
+      </a>
+
+      <!-- Divider -->
+      <hr class="sidebar-divider my-0">
+
+      <!-- Nav Item -->
+      <li class="nav-item active">
+        <a class="nav-link" th:href="@{/}">
+          <i class="fas fa-table"></i>
+          <span>社員一覧</span>
+        </a>
+      </li>
+      <li class="nav-item active">
+        <a class="nav-link" th:href="@{/add}">
+          <i class="fas fa-user-plus"></i>
+          <span>社員登録</span>
+        </a>
+      </li>
+
+      <!-- Divider -->
+      <hr class="sidebar-divider d-none d-md-block">
+
+      <!-- Sidebar Toggler (Sidebar) -->
+      <div class="text-center d-none d-md-inline">
+        <button class="rounded-circle border-0"
+          id="sidebarToggle">
+        </button>
+      </div>
+
+    </ul>
+    <!-- End of Sidebar -->
+  </th:block>
+</body>
+</html>
